# generated/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_set_flag_api_v1_admin_flags__code__post": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "Body_set_flag_api_v1_admin_flags__code__post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Governed catalog of data products and the AI agents that run on them.\n\nEvery response that carries a computed number also carries the rubric version it\nwas computed under, so a score, a ranking or a value figure can be replayed\nexactly. Every 403 names the missing scope and links to a pre-filled access\nrequest. An agent answer that cannot be grounded is withheld, never guessed.",
    "title": "Data Product and AI Agent Marketplace API",
    "version": "0.1.0",
    "x-generated": {
      "by": "scripts/gen.py",
      "from": "services/api/main.py",
      "generator_version": "1.0.0",
      "manifest_hash": "050a8e25b3257929c47ba51039b44f34fd5d60a2d9d9f41eb4b3a53e07995239",
      "warning": "DO NOT EDIT"
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/connectors": {
      "get": {
        "operationId": "connectors_api_v1_admin_connectors_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-marketplace-subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Subject"
            }
          },
          {
            "in": "header",
            "name": "x-marketplace-agent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Connectors Api V1 Admin Connectors Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source systems",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/flags": {
      "get": {
        "operationId": "flags_api_v1_admin_flags_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-marketplace-subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Subject"
            }
          },
          {
            "in": "header",
            "name": "x-marketplace-agent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Flags Api V1 Admin Flags Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feature flags",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/flags/{code}": {
      "post": {
        "operationId": "set_flag_api_v1_admin_flags__code__post",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-marketplace-subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Subject"
            }
          },
          {
            "in": "header",
            "name": "x-marketplace-agent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Agent"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_set_flag_api_v1_admin_flags__code__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Set Flag Api V1 Admin Flags  Code  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle a flag",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/rubrics": {
      "get": {
        "operationId": "list_rubrics_api_v1_admin_rubrics_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-marketplace-subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Subject"
            }
          },
          {
            "in": "header",
            "name": "x-marketplace-agent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Rubrics Api V1 Admin Rubrics Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Every rubric",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/rubrics/{code}": {
      "get": {
        "operationId": "read_rubric_api_v1_admin_rubrics__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-marketplace-subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Subject"
            }
          },
          {
            "in": "header",
            "name": "x-marketplace-agent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Read Rubric Api V1 Admin Rubrics  Code  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One rubric and its history",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/rubrics/{code}/versions": {
      "post": {
        "description": "Publish a new version, then re-score under it.\n\nThe acceptance criterion for M12.2 is that a weight change here re-scores\nthe estate *and* preserves prior snapshots, so both happen in one call and\nthe response reports both counts. A console that published without\nre-scoring would leave the estate describing itself under a rubric no longer\nin force.",
        "operationId": "publish_version_api_v1_admin_rubrics__code__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-marketplace-subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Subject"
            }
          },
          {
            "in": "header",
            "name": "x-marketplace-agent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Agent"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Publish Version Api V1 Admin Rubrics  Code  Versions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish an edited rubric",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/taxonomies": {
      "get": {
        "operationId": "taxonomies_api_v1_admin_taxonomies_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-marketplace-subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Subject"
            }
          },
          {
            "in": "header",
            "name": "x-marketplace-agent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Taxonomies Api V1 Admin Taxonomies Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Controlled vocabulary",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/tenancy": {
      "get": {
        "operationId": "tenancy_api_v1_admin_tenancy_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-marketplace-subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Subject"
            }
          },
          {
            "in": "header",
            "name": "x-marketplace-agent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Marketplace-Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Tenancy Api V1 Admin Tenancy Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tenancy and isolation",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}